requirements.txt
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
pydantic==1.8.2
pyodbc==4.0.30
python-dotenv==0.17.1
SQLAlchemy==1.4.17
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
